move filtering functions to filter.py
