+++ stewi/RCRAInfo.py
@@ -107,9 +107,10 @@ import time, datetime
 from stewi.globals import write_metadata, data_dir, config,\
     checkforFile, USton_kg, get_reliability_table_for_source, paths,\
     log, store_inventory, compile_source_metadata, read_source_metadata,\
-    filter_states, aggregate, create_paths_if_missing, set_stewi_meta
+    aggregate, create_paths_if_missing, set_stewi_meta
 from stewi.validate import update_validationsets_sources, validate_inventory,\
     write_validation_result
+from stewi.validate import filter_states
 
 try:
     from selenium import webdriver
